Rename Holder to Container, broke handler.go into various files
Contextual handler.Container is easier to read and understand than
handler.Holder and to make it easier to change the overall
messages/handler structure the one huge handler.go is separated into:

- messages/handler/container.go - containing the container definition
- messages/handler/handler.go - containing the interfaces
- messages/handler/on-message.go - containing the on message impl
